Updated tinker documentation to read better
There were several portions of this document that could use some clarification. Please double check my explanation of `AddWebsiteFilterOnCommand` in the following paragraph. I attempted my best understanding as to why it was not a good option to utilizing in extending Tinker.

```You can only tinker on one database at a time, therefore, the `website_id` needs to be a required option. Utilizing the `AddWebsiteFilterOnCommand` trait available in the Tenancy package would not work in this instance; it is tailored to run processes on multiple websites and makes the `website_id` parameter optional.```