Add clarity to installing Spatie Permissions
Spatie Permissions will create the migration to create the permissions table in the `migrations` directory, but the example that follows uses the tenant connection in the model. The model will expect the tables to be in each tenant database, which it won't be given where Spatie Permissions placed the migration by default.

The example also suggests to add `UsesTenantConnection` to the `User` model, but doing so will also require the user table migration to be in the `migrations/tenant` directory. Based on the documentation, the default will be for the `User` model to `UseSystemConnection` and the permissions to be at a tenant level.